Refactor: new ShowObject catalog hierarchy structure (#273)
* Refactor: new ShowObject catalog hierarchy structure

* Move functional AttributeBox to new object views dir
