+++ mwdb/web/src/components/ShowObject/Actions/RemoveAction.js
