+++ mwdb/web/src/components/ShowObject/Views/LatestConfigTab.js
@@ -1,6 +1,6 @@
 import React, { useContext } from "react";
 
-import ConfigTable from "../ConfigTable";
+import ConfigTable from "../../ConfigTable";
 
 import { ObjectContext } from "@mwdb-web/commons/context";
 import { ObjectAction, ObjectTab } from "@mwdb-web/commons/ui";
